Split an "mm" feature out of the "io" module.
