+++ src/imp/linux_raw/io/syscalls.rs
@@ -15,26 +15,25 @@
 use super::super::arch::choose::syscall1;
 use super::super::arch::choose::{
     syscall1_readonly, syscall2, syscall2_readonly, syscall3, syscall3_readonly, syscall4,
-    syscall4_readonly, syscall5, syscall5_readonly, syscall6,
+    syscall4_readonly, syscall5, syscall5_readonly,
 };
 use super::super::c;
+#[cfg(target_pointer_width = "64")]
+use super::super::conv::loff_t_from_u64;
 use super::super::conv::{
-    borrowed_fd, by_ref, c_int, c_uint, no_fd, out, pass_usize, raw_fd, ret, ret_discarded_fd,
-    ret_owned_fd, ret_usize, ret_void_star, slice, slice_mut, void_star, zero,
+    borrowed_fd, by_ref, c_int, c_uint, out, pass_usize, raw_fd, ret, ret_discarded_fd,
+    ret_owned_fd, ret_usize, slice, slice_mut, void_star, zero,
 };
 use super::super::reg::nr;
 use crate::fd::{AsFd, BorrowedFd, RawFd};
 use crate::io::{
-    self, epoll, Advice, DupFlags, EventfdFlags, IoSlice, IoSliceMut, MapFlags, MlockFlags,
-    MprotectFlags, MremapFlags, MsyncFlags, OwnedFd, PipeFlags, PollFd, ProtFlags, ReadWriteFlags,
-    UserfaultfdFlags,
+    self, epoll, DupFlags, EventfdFlags, IoSlice, IoSliceMut, OwnedFd, PipeFlags, PollFd,
+    ReadWriteFlags,
 };
 #[cfg(feature = "net")]
 use crate::net::{RecvFlags, SendFlags};
 use core::cmp;
 use core::mem::MaybeUninit;
-#[cfg(any(target_arch = "arm", target_arch = "mips", target_arch = "x86"))]
-use linux_raw_sys::general::__NR_mmap2;
 #[cfg(not(any(
     target_arch = "aarch64",
     target_arch = "mips",
@@ -44,20 +43,16 @@ use linux_raw_sys::general::__NR_mmap2;
 use linux_raw_sys::general::__NR_pipe;
 use linux_raw_sys::general::{
     __NR_close, __NR_dup, __NR_dup3, __NR_epoll_create1, __NR_epoll_ctl, __NR_eventfd2, __NR_ioctl,
-    __NR_madvise, __NR_mlock, __NR_mlock2, __NR_mprotect, __NR_mremap, __NR_msync, __NR_munlock,
-    __NR_munmap, __NR_pipe2, __NR_pread64, __NR_preadv, __NR_preadv2, __NR_pwrite64, __NR_pwritev,
-    __NR_pwritev2, __NR_read, __NR_readv, __NR_userfaultfd, __NR_write, __NR_writev, epoll_event,
-    EPOLL_CTL_ADD, EPOLL_CTL_DEL, EPOLL_CTL_MOD,
+    __NR_pipe2, __NR_pread64, __NR_preadv, __NR_preadv2, __NR_pwrite64, __NR_pwritev,
+    __NR_pwritev2, __NR_read, __NR_readv, __NR_write, __NR_writev, epoll_event, EPOLL_CTL_ADD,
+    EPOLL_CTL_DEL, EPOLL_CTL_MOD,
 };
 use linux_raw_sys::ioctl::{BLKPBSZGET, BLKSSZGET, FIONBIO, FIONREAD, TIOCEXCL, TIOCNXCL};
 #[cfg(target_pointer_width = "32")]
 use {
-    super::super::arch::choose::syscall6_readonly,
+    super::super::arch::choose::{syscall6, syscall6_readonly},
     super::super::conv::{hi, lo},
-    core::convert::TryInto,
 };
-#[cfg(target_pointer_width = "64")]
-use {super::super::conv::loff_t_from_u64, linux_raw_sys::general::__NR_mmap};
 #[cfg(any(target_arch = "aarch64", target_arch = "riscv64"))]
 use {
     super::super::conv::{opt_ref, size_of},
@@ -355,28 +350,6 @@ pub(crate) unsafe fn close(fd: RawFd) {
 }
 
 #[inline]
-pub(crate) fn madvise(addr: *mut c::c_void, len: usize, advice: Advice) -> io::Result<()> {
-    unsafe {
-        ret(syscall3(
-            nr(__NR_madvise),
-            void_star(addr),
-            pass_usize(len),
-            c_uint(advice as c::c_uint),
-        ))
-    }
-}
-
-#[inline]
-pub(crate) unsafe fn msync(addr: *mut c::c_void, len: usize, flags: MsyncFlags) -> io::Result<()> {
-    ret(syscall3(
-        nr(__NR_msync),
-        void_star(addr),
-        pass_usize(len),
-        c_uint(flags.bits()),
-    ))
-}
-
-#[inline]
 pub(crate) fn eventfd(initval: u32, flags: EventfdFlags) -> io::Result<OwnedFd> {
     unsafe {
         ret_owned_fd(syscall2_readonly(
@@ -547,199 +520,6 @@ pub(crate) fn dup3(fd: BorrowedFd<'_>, new: &OwnedFd, flags: DupFlags) -> io::Re
     }
 }
 
-/// # Safety
-///
-/// `mmap` is primarily unsafe due to the `addr` parameter, as anything working
-/// with memory pointed to by raw pointers is unsafe.
-#[inline]
-pub(crate) unsafe fn mmap(
-    addr: *mut c::c_void,
-    length: usize,
-    prot: ProtFlags,
-    flags: MapFlags,
-    fd: BorrowedFd<'_>,
-    offset: u64,
-) -> io::Result<*mut c::c_void> {
-    #[cfg(target_pointer_width = "32")]
-    {
-        ret_void_star(syscall6(
-            nr(__NR_mmap2),
-            void_star(addr),
-            pass_usize(length),
-            c_uint(prot.bits()),
-            c_uint(flags.bits()),
-            borrowed_fd(fd),
-            (offset / 4096)
-                .try_into()
-                .map(|scaled_offset| pass_usize(scaled_offset))
-                .map_err(|_| io::Error::INVAL)?,
-        ))
-    }
-    #[cfg(target_pointer_width = "64")]
-    {
-        ret_void_star(syscall6(
-            nr(__NR_mmap),
-            void_star(addr),
-            pass_usize(length),
-            c_uint(prot.bits()),
-            c_uint(flags.bits()),
-            borrowed_fd(fd),
-            loff_t_from_u64(offset),
-        ))
-    }
-}
-
-/// # Safety
-///
-/// `mmap` is primarily unsafe due to the `addr` parameter, as anything working
-/// with memory pointed to by raw pointers is unsafe.
-#[inline]
-pub(crate) unsafe fn mmap_anonymous(
-    addr: *mut c::c_void,
-    length: usize,
-    prot: ProtFlags,
-    flags: MapFlags,
-) -> io::Result<*mut c::c_void> {
-    #[cfg(target_pointer_width = "32")]
-    {
-        ret_void_star(syscall6(
-            nr(__NR_mmap2),
-            void_star(addr),
-            pass_usize(length),
-            c_uint(prot.bits()),
-            c_uint(flags.bits() | linux_raw_sys::general::MAP_ANONYMOUS),
-            no_fd(),
-            pass_usize(0),
-        ))
-    }
-    #[cfg(target_pointer_width = "64")]
-    {
-        ret_void_star(syscall6(
-            nr(__NR_mmap),
-            void_star(addr),
-            pass_usize(length),
-            c_uint(prot.bits()),
-            c_uint(flags.bits() | linux_raw_sys::general::MAP_ANONYMOUS),
-            no_fd(),
-            loff_t_from_u64(0),
-        ))
-    }
-}
-
-#[inline]
-pub(crate) unsafe fn mprotect(
-    ptr: *mut c::c_void,
-    len: usize,
-    flags: MprotectFlags,
-) -> io::Result<()> {
-    ret(syscall3(
-        nr(__NR_mprotect),
-        void_star(ptr),
-        pass_usize(len),
-        c_uint(flags.bits()),
-    ))
-}
-
-/// # Safety
-///
-/// `munmap` is primarily unsafe due to the `addr` parameter, as anything
-/// working with memory pointed to by raw pointers is unsafe.
-#[inline]
-pub(crate) unsafe fn munmap(addr: *mut c::c_void, length: usize) -> io::Result<()> {
-    ret(syscall2(
-        nr(__NR_munmap),
-        void_star(addr),
-        pass_usize(length),
-    ))
-}
-
-/// # Safety
-///
-/// `mremap` is primarily unsafe due to the `old_address` parameter, as
-/// anything working with memory pointed to by raw pointers is unsafe.
-#[inline]
-pub(crate) unsafe fn mremap(
-    old_address: *mut c::c_void,
-    old_size: usize,
-    new_size: usize,
-    flags: MremapFlags,
-) -> io::Result<*mut c::c_void> {
-    ret_void_star(syscall4(
-        nr(__NR_mremap),
-        void_star(old_address),
-        pass_usize(old_size),
-        pass_usize(new_size),
-        c_uint(flags.bits()),
-    ))
-}
-
-/// # Safety
-///
-/// `mremap_fixed` is primarily unsafe due to the `old_address` and
-/// `new_address` parameters, as anything working with memory pointed to by raw
-/// pointers is unsafe.
-#[inline]
-pub(crate) unsafe fn mremap_fixed(
-    old_address: *mut c::c_void,
-    old_size: usize,
-    new_size: usize,
-    flags: MremapFlags,
-    new_address: *mut c::c_void,
-) -> io::Result<*mut c::c_void> {
-    ret_void_star(syscall5(
-        nr(__NR_mremap),
-        void_star(old_address),
-        pass_usize(old_size),
-        pass_usize(new_size),
-        c_uint(flags.bits()),
-        void_star(new_address),
-    ))
-}
-
-/// # Safety
-///
-/// `mlock` operates on raw pointers and may round out to the nearest page
-/// boundaries.
-#[inline]
-pub(crate) unsafe fn mlock(addr: *mut c::c_void, length: usize) -> io::Result<()> {
-    ret(syscall2(
-        nr(__NR_mlock),
-        void_star(addr),
-        pass_usize(length),
-    ))
-}
-
-/// # Safety
-///
-/// `mlock_with` operates on raw pointers and may round out to the nearest page
-/// boundaries.
-#[inline]
-pub(crate) unsafe fn mlock_with(
-    addr: *mut c::c_void,
-    length: usize,
-    flags: MlockFlags,
-) -> io::Result<()> {
-    ret(syscall3(
-        nr(__NR_mlock2),
-        void_star(addr),
-        pass_usize(length),
-        c_uint(flags.bits()),
-    ))
-}
-
-/// # Safety
-///
-/// `munlock` operates on raw pointers and may round out to the nearest page
-/// boundaries.
-#[inline]
-pub(crate) unsafe fn munlock(addr: *mut c::c_void, length: usize) -> io::Result<()> {
-    ret(syscall2(
-        nr(__NR_munlock),
-        void_star(addr),
-        pass_usize(length),
-    ))
-}
-
 #[inline]
 pub(crate) fn pipe_with(flags: PipeFlags) -> io::Result<(OwnedFd, OwnedFd)> {
     unsafe {
@@ -814,14 +594,6 @@ pub(crate) fn poll(fds: &mut [PollFd<'_>], timeout: c::c_int) -> io::Result<usiz
             c_int(timeout),
         ))
     }
-}
-
-#[inline]
-pub(crate) unsafe fn userfaultfd(flags: UserfaultfdFlags) -> io::Result<OwnedFd> {
-    ret_owned_fd(syscall1_readonly(
-        nr(__NR_userfaultfd),
-        c_uint(flags.bits()),
-    ))
 }
 
 #[inline]
